# 4_backend/4_3_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API for working with users and their roles.",
    "description": "RESTful API for managing users and roles",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/create_user": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Creating a user.",
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful user creation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Responce"
                },
                "example": {
                  "Success": "User created successfully."
                }
              }
            }
          },
          "201": {
            "description": "Error creating user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Responce"
                },
                "example": {
                  "Error": "Duplicate user email."
                }
              }
            }
          }
        }
      }
    },
    "/api/delete_user/{id_user}": {
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Deleting a user.",
        "operationId": "delete_user",
        "parameters": [
          {
            "name": "id_user",
            "in": "path",
            "description": "User ID to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Responce"
                },
                "example": {
                  "Success": "The user has been deleted."
                }
              }
            }
          },
          "201": {
            "description": "Error deleting user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Responce"
                },
                "example": {
                  "Error": "The user does not exist."
                }
              }
            }
          }
        }
      }
    },
    "/api/initdb": {
      "get": {
        "tags": [
          "initdb"
        ],
        "summary": "Creating database objects.",
        "operationId": "initdb_handle",
        "responses": {
          "200": {
            "description": "Successfully created database objects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Responce"
                },
                "example": {
                  "Success": "Database objects created successfully."
                }
              }
            }
          },
          "201": {
            "description": "Error creating database objects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Responce"
                },
                "example": {
                  "Error": "Error creating trigger."
                }
              }
            }
          }
        }
      }
    },
    "/api/update_useremail": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update user's email.",
        "operationId": "update_useremail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmailUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful modification of user email.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Responce"
                },
                "example": {
                  "Success": "User's email changed successfully."
                }
              }
            }
          },
          "201": {
            "description": "Error modifying user email.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Responce"
                },
                "example": {
                  "Error": "Invalid email: n1#abc.com"
                }
              }
            }
          }
        }
      }
    },
    "/api/update_username": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update user's username",
        "operationId": "update_username",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNameUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful modification of username.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Responce"
                },
                "example": {
                  "Success": "User's name changed successfully."
                }
              }
            }
          },
          "201": {
            "description": "Error modifying username.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Responce"
                },
                "example": {
                  "Error": "Not found user for id_user: 100"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateUser": {
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "Responce": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "Success"
            ],
            "properties": {
              "Success": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "Error"
            ],
            "properties": {
              "Error": {
                "type": "string"
              }
            }
          }
        ]
      },
      "UpdateEmailUser": {
        "type": "object",
        "description": "Модифицировать email у пользователя",
        "required": [
          "new_email",
          "id_user"
        ],
        "properties": {
          "id_user": {
            "type": "integer",
            "format": "int32",
            "description": "Id user",
            "minimum": 0
          },
          "new_email": {
            "type": "string"
          }
        }
      },
      "UpdateNameUser": {
        "type": "object",
        "description": "Модифицировать имя у пользователю",
        "required": [
          "new_name",
          "id_user"
        ],
        "properties": {
          "id_user": {
            "type": "integer",
            "format": "int32",
            "description": "Id user",
            "minimum": 0
          },
          "new_name": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "initdb",
      "description": "Creating the necessary objects in the database"
    },
    {
      "name": "users",
      "description": "Working with users"
    }
  ]
}