# 4_backend/4_3_api/src/server_executor.rs
use std::{
        //path::Path, 
        sync::Arc
    };

use const_format::concatcp;

use crate::{args::{UpdateEmailUser, UpdateNameUser}, common::{self, Responce}, roles, users::{self, User}, users_roles} ;

//use axum::extract::Path;
use axum::{
        extract::{
            self, Json, State
        }, http::{StatusCode, status}, response::IntoResponse
} ;

use anyhow::Result ;
use clap::Command;

use crate::db::Database ;

use crate::args ;

use utoipa::{OpenApi, ToSchema} ;


const DB_OBJ_CREATED_SUCCESS: &str = "Database objects created successfully." ;

/// Файл с данными документации
const OPENAPI_FILE: &str = "openapi.json" ;

/// url документации openapi
pub const OPENAPI_URL_DOCS: &str = "/docs" ;

/// url спецификации openapi
pub const OPENAPI_URL_SPECIFIC: &str = "/api-docs/openapi.json" ;

/// пользователь успешно создан
const USER_CREATED_SUCCESSFULY: &str = "User created successfully." ;

/// тэг пользователи
const TAG_USERS: &str = "users" ;

/// id_user ключ
pub const ID_USER_KEY: &str = "id_user" ;

/// Сообщение имя пользователя изменено успешно
const USER_NAME_CHANGED_SUCCESS: &str = "User's name changed successfully." ;

/// Сообщение email пользователя изменено успешно
const USER_EMAIL_CHANGED_SUCCESS: &str = "User's email changed successfully." ;

/// Запись в файл спецификации openapi если спецификация изменилась
pub fn write_to_openapi(op_api: &utoipa::openapi::OpenApi) ->Result<()> {

    // текущий контент сожержимого openapi
    let content_openapi_now = serde_json::to_string_pretty(&op_api)? ;

    // путь к файлу openapi
    let openapi_path = std::path::Path::new(OPENAPI_FILE) ;

    // чтение содержимого файла openapi
    let content_openapi = if openapi_path.exists() {
        std::fs::read_to_string(openapi_path)?
    } else {
        "".to_string()
    } ;

    if content_openapi_now != content_openapi {
        std::fs::write(
                openapi_path,
                content_openapi_now
            )? ;
    }

    Ok(())
}

// Отправка сообщения об успехе
fn success_message(mess: &str) -> common::Responce {
    common::Responce::Success(mess.to_owned())
}

// Отправка сообщения об ошибке
fn error_message(err: &str) ->common::Responce {
    common::Responce::Error(err.to_owned())
}

// OpenAPI документация
#[derive(OpenApi)]
#[openapi(
    paths(
        initdb_handle,
        create_user,
        delete_user,
        update_username,
        update_useremail,
    ),
    components(
        schemas(
            common::Responce,
            args::CreateUser,
            args::UpdateNameUser,
            args::UpdateEmailUser,
        )
    ),
    tags(
        (
            name = "initdb",
            description = "Creating the necessary objects in the database",
        ),
        (
            name = "users",
            description = "Working with users",
        ),        
    ),
    info(
        title = "API for working with users and their roles.",
        version = "1.0.0",
        description = "RESTful API for managing users and roles",
    ),
  )
 ]
pub struct ApiDoc;

// ****************** Раздел INIT-DB ******************

/// инициализация базы данных реальная внутрянная функция
async fn initdb_handle_int(
                db_res: &Database,
            ) ->Result<common::Responce>{
    db_res
        .create_tables()
        .await? ;

    Ok(success_message(
            DB_OBJ_CREATED_SUCCESS
        ))
}

// инициализация базы данных
#[utoipa::path(
    get,
    path = concatcp!(
                common::BASE_URI_PATH,        // "/api/",
                common::INIT_DB_PART      // "initdb"
            ),  // "/api/initdb"
    summary = "Creating database objects.",
    responses (
        (
            status = StatusCode::OK,  // 200, 
            description = "Successfully created database objects.", 
            body = common::Responce,
            example = json!({"Success": DB_OBJ_CREATED_SUCCESS})
        ),
        (
            status = StatusCode::CREATED,   // 303, 
            description = "Error creating database objects.", 
            body = common::Responce,
            example = json!({"Error": "Error creating trigger."})
        ),
    ),
    tag = "initdb",
  )
]
pub async fn initdb_handle(
                State(db_res): State<Arc<Database>>,
            ) 
                //->Json<common::Responce> 
                -> impl IntoResponse
            {

    match initdb_handle_int(&db_res)
                .await 
    {
        Ok(v) => (
                            StatusCode::OK,
                            Json(v)
                        ),
        Err(err) => (
                            StatusCode::CREATED,
                            Json(
                                error_message(
                                    &err.to_string()
                                )
                            ),
                        ),
    }
}

// ****************** Раздел User ******************

// создать пользователя,внутренний формат
async fn create_user_int(db_res: &Database,
                        cmd: &args::CreateUser,
                        ) ->Result<common::Responce> {
    let mut trans = db_res
                      .pool
                      .begin()
                      .await? ;

    let new_user = 
            users::User::ins_user(
                    &mut *trans,
                    &cmd.name, //        name, 
                    &cmd.email, //    email
                )
                .await? ;

    users_roles::UsersRoles::ins_role_to_user(
            &mut *trans, 
            new_user.id_user(),
            roles::SLUG_DEFAULT
        )
        .await? ;
        
    trans.commit().await? ;

    Ok(success_message(USER_CREATED_SUCCESSFULY))
}


// создать пользователя
#[utoipa::path(
    post,
    path = concatcp!(
                common::BASE_URI_PATH,        // "/api/",
                common::CREATE_USER_PART      // "create_user"
            ), // "/api/create_user",
    summary = "Creating a user.",
    request_body = args::CreateUser,
    responses (
        (
            status = StatusCode::OK,  // 200, 
            description = "Successful user creation.", 
            body = common::Responce,
            example = json!({"Success": "User created successfully."})
        ),
        (
            status = StatusCode::CREATED,   // 303, 
            description = "Error creating user.", 
            body = common::Responce,
            example = json!({"Error": "Duplicate user email."})
        ),
    ),
    tag = TAG_USERS,
  )
]
pub async fn create_user(
                    State(db_res): State<Arc<Database>>,
                    Json(cmd): Json<args::CreateUser>
                ) ->impl IntoResponse {

    match create_user_int(&db_res, &cmd).await {
            Ok(v) => (
                            StatusCode::OK,
                            Json(v)
                           ),
            Err(err) => (
                            StatusCode::CREATED,
                            Json(
                                error_message(&err.to_string())
                              )
                            ),
    }
}

// удалить пользователя, внутренний формат
async fn delete_user_int(
            db_res:     &Database,
            id_user:    u32,
         ) ->Result<common::Responce> {
    
    let mut tmp_user = User::default() ;

    tmp_user.set_id_user(id_user)? ;

    // Сформировать новую транзакцию
    let mut trans = 
                db_res
                    .pool
                    // Устанавливает соединение и немедленно начинает новую транзакцию.
                    .begin()
                    .await? ;

    // удалить пользователя
    User::delete_user(&mut *trans, tmp_user.id_user()).await? ;
    
    // выполнить commit в DB
    trans.commit().await? ;    

    Ok(success_message("The user has been deleted."))
}

// удалить пользователя
#[utoipa::path(
    delete,
    //path =  &format!("{}{{id_user}}", common::get_delete_user_uri_short()), // "/api/del_user/{id_user}",
    path =  &format!("{}{{{}}}", common::get_delete_user_uri_short(), ID_USER_KEY), // "/api/del_user/{id_user}",
    summary = "Deleting a user.",
    params(
        ("id_user" = u32, Path, description = "User ID to delete")
    ),
    responses(
        (
            status = StatusCode::OK, // 200, 
            description = "Successfully deleted user.", 
            body = common::Responce,
            example = json!({"Success": "The user has been deleted."}),
        ),
        (
            status = StatusCode::CREATED,  // 303,
            description = "Error deleting user.",
            body = common::Responce,
            example = json!({"Error": "The user does not exist."}),
        ),
    ),
    tag = TAG_USERS,
    /*
    operation_id = "delete_user",
    security(
        ("bearer_auth" = [])
    )
     */
)]
pub async fn delete_user(
                  State(db_res): State<Arc<Database>>,
                  extract::Path(id_user): extract::Path<u32>,
                ) ->impl IntoResponse {
    //println!("id_user: {}", id_user) ;
    match delete_user_int(&db_res, id_user).await {
        Ok(v) => (
                            StatusCode::OK,
                            Json(v)
                           ),
        Err(err) => (
                        StatusCode::CREATED,
                        Json(error_message(&err.to_string()))
                    ),
    }
}

// модифицировать имя пользователя, внутренний формат
async fn update_username_int(
            db_res:     &Database,
            com:        &UpdateNameUser,
         ) ->Result<common::Responce>{
    
    // Сформировать новую транзакцию
    let mut trans = 
              db_res
                .pool
                // Устанавливает соединение и немедленно начинает новую транзакцию.
                .begin()
                .await? ;

    // Модифицировать имя пользователя
    users::User::update_name(
                    &mut *trans,
                    &com.new_name,
                    com.id_user
                ).await? ;

    // Выполнить commit
    trans.commit().await? ;

    Ok(success_message(USER_NAME_CHANGED_SUCCESS))
}

// модифицировать имя пользователя
#[utoipa::path(
    put,
    path = concatcp!(
                common::BASE_URI_PATH,        // "/api/",
                common::UPDATE_USERNAME_PART      // "create_user"
            ), // "/api/create_user",
    summary = "Update user's username",
    request_body = args::UpdateNameUser,
    responses (
        (
            status = StatusCode::OK,  // 200, 
            description = "Successful modification of username.", 
            body = common::Responce,
            example = json!({"Success": USER_NAME_CHANGED_SUCCESS})
        ),
        (
            status = StatusCode::CREATED,   // 303, 
            description = "Error modifying username.", 
            body = common::Responce,
            example = json!({"Error": "Not found user for id_user: 100"})
        ),
    ),
    tag = TAG_USERS,
  )
]
pub async fn update_username(
                State(db_res): State<Arc<Database>>,
                Json(com): Json<args::UpdateNameUser> 
             ) ->impl IntoResponse {
    match update_username_int(&db_res, &com).await {
        Ok(v) => (
                            StatusCode::OK,
                            Json(v),
                           ),
        Err(err) => (
                            StatusCode::CREATED,
                            Json(
                                error_message(
                                    &err.to_string()
                                )
                            )
                           ),
    }
}

// модифицировать email пользователя, внутренний формат
async fn update_useremail_int(
                    db_res:     &Database,
                    com:  &UpdateEmailUser
                ) ->Result<common::Responce> {
    // Сформировать новую транзакцию
    let mut trans = 
               db_res
                .pool
                // Устанавливает соединение и немедленно начинает новую транзакцию.
                .begin()
                .await? ;

    // Модифицировать имя у пользователя
    users::User::update_email(
            &mut *trans,
            &com.new_email,
            com.id_user
          )
          .await? ;

    // Выполнить commit
    trans.commit().await? ;

    Ok(success_message(USER_EMAIL_CHANGED_SUCCESS))
}

// модифицировать email пользователя
#[utoipa::path(
    put,
    path = concatcp!(
                common::BASE_URI_PATH,        // "/api/",
                common::UPDATE_USEREMAIL_PART
            ), // "/api/create_user",
    summary = "Update user's email.",
    request_body = args::UpdateEmailUser,
    responses (
        (
            status = StatusCode::OK,  // 200, 
            description = "Successful modification of user email.", 
            body = common::Responce,
            example = json!({"Success": USER_EMAIL_CHANGED_SUCCESS})
        ),
        (
            status = StatusCode::CREATED,   // 303, 
            description = "Error modifying user email.", 
            body = common::Responce,
            example = json!({"Error": "Invalid email: n1#abc.com"})
        ),
    ),
    tag = TAG_USERS,
  )
]
pub async fn update_useremail(
                    State(db_res): State<Arc<Database>>,
                    Json(com):  Json<args::UpdateEmailUser>
                ) ->impl IntoResponse {
    match update_useremail_int(&db_res, &com).await {
        Ok(v) => (
                            StatusCode::OK,
                            Json(v)
                        ),
        Err(err) => (
                            StatusCode::CREATED,
                            Json(
                                error_message(&err.to_string())
                            )
                        )
    }
}